move dialog-info & dialog-board-game into activities folder
