+++ src/app/activities/dialog-board-game/dialog-board-game.component.css
