+++ src/app/activities/dialog-info/dialog-info.component.css
